+++ src/nvim/textobject.h
@@ -1,0 +1,11 @@
+#ifndef NVIM_TEXTOBJECT_H
+#define NVIM_TEXTOBJECT_H
+
+#include "nvim/normal.h"  // for oparg_T
+#include "nvim/pos.h"  // for linenr_T
+#include "nvim/vim.h"  // for Direction
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "textobject.h.generated.h"
+#endif
+#endif  // NVIM_TEXTOBJECT_H
